// src/api/generated/models/dataScope.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Arco Enterprise Admin API
 * Spring Boot 管理后台示例契约，建议由 springdoc-openapi 的 /v3/api-docs 替换。
 * OpenAPI spec version: 0.1.0
 */

export type DataScope = typeof DataScope[keyof typeof DataScope];


export const DataScope = {
  all: 'all',
  department: 'department',
  self: 'self',
} as const;